// app/model/dto/PersonaDTO.ts
export interface PersonaDTO {
    id?: number | null;
    nombre?: string | null;
    altura?: number | null;
    peso?: number | null;
    color_cabello?: string | null;
    color_piel?: string | null;
    color_ojos?: string | null;
    anio_nacimiento?: string | null;
    genero?: string | null;
}